refactor(ui/repos): hoist groupByAgent to lib for testability
Addresses review feedback on PR #254: the agent-bucketing logic was
defined inline inside the repos-page render callback, so it was
recreated on every render and unreachable by unit tests. Moved it to
`src/lib/bindings.ts` alongside the shared `Binding` interface.

Test infrastructure (vitest) arrives with PR #247; unit tests for
`groupByAgent` (insertion-order preservation, multi-binding grouping,
empty-input) will follow once that merges.
